// src/app/interfaces/reporte-mensual.ts
export interface ReporteMensual {
    cnt: number;
    name: string;
    fecha: string;
    hroIngreso: string;
    hroSalida: string;
    hroMarcadoIngreso: string;
    hroMarcadoSalida: string;
    dia: string;
    tipo: string;
    materia: string;
    minutoAtraso: string;
    semestre: string;
    turno: string;
}